# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A concise Gin backend solution with config, db, tracing, metrics and graceful shutdown.",
        "title": "Ops Container API",
        "contact": {
            "name": "Liu Haogui"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/ops/ping": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ops"
                ],
                "summary": "Ping",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.StringDataResponse"
                        }
                    }
                }
            }
        },
        "/api/ops/version": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ops"
                ],
                "summary": "Get service version",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.VersionDataResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/config": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Read current runtime config",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/config.Config"
                        }
                    }
                }
            }
        },
        "/api/v1/container/log/{id}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "container"
                ],
                "summary": "Get container logs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Container ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Number of log lines",
                        "name": "lines",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.StringListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.StringDataResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.StringDataResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.StringDataResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/container/start/{id}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "container"
                ],
                "summary": "Start a container",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Container ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.StringDataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.StringDataResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.StringDataResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.StringDataResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/container/stop/{id}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "container"
                ],
                "summary": "Stop a container",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Container ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.StringDataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.StringDataResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.StringDataResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.StringDataResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/container/tail-log/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "container"
                ],
                "summary": "Tail container logs over websocket",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Container ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "101": {
                        "description": "Switching Protocols",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.StringDataResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.StringDataResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/container/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "container"
                ],
                "summary": "Get container list by instance id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Instance ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "1: instance containers, 2: init containers, 3: other containers",
                        "name": "container_type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.StringDataResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.StringDataResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.StringDataResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/health": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.HealthResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "config.AppConfig": {
            "type": "object",
            "properties": {
                "env": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "config.AuthConfig": {
            "type": "object",
            "properties": {
                "tokens": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "config.Config": {
            "type": "object",
            "properties": {
                "app": {
                    "$ref": "#/definitions/config.AppConfig"
                },
                "auth": {
                    "$ref": "#/definitions/config.AuthConfig"
                },
                "database": {
                    "$ref": "#/definitions/config.DatabaseConfig"
                },
                "development": {
                    "$ref": "#/definitions/config.DevelopmentConfig"
                },
                "log": {
                    "$ref": "#/definitions/config.LogConfig"
                },
                "prometheus": {
                    "$ref": "#/definitions/config.PrometheusConfig"
                },
                "server": {
                    "$ref": "#/definitions/config.ServerConfig"
                },
                "swagger": {
                    "$ref": "#/definitions/config.SwaggerConfig"
                },
                "telemetry": {
                    "$ref": "#/definitions/config.TelemetryConfig"
                }
            }
        },
        "config.DatabaseConfig": {
            "type": "object",
            "properties": {
                "connMaxLifetime": {
                    "type": "integer"
                },
                "driver": {
                    "type": "string"
                },
                "dsn": {
                    "type": "string"
                },
                "enable": {
                    "type": "boolean"
                },
                "maxIdleConns": {
                    "type": "integer"
                },
                "maxOpenConns": {
                    "type": "integer"
                }
            }
        },
        "config.DevelopmentConfig": {
            "type": "object",
            "properties": {
                "enablePProf": {
                    "type": "boolean"
                }
            }
        },
        "config.LogConfig": {
            "type": "object",
            "properties": {
                "compress": {
                    "type": "boolean"
                },
                "filename": {
                    "type": "string"
                },
                "format": {
                    "type": "string"
                },
                "level": {
                    "type": "string"
                },
                "maxAgeDays": {
                    "type": "integer"
                },
                "maxBackups": {
                    "type": "integer"
                },
                "maxSizeMB": {
                    "type": "integer"
                }
            }
        },
        "config.PrometheusConfig": {
            "type": "object",
            "properties": {
                "enable": {
                    "type": "boolean"
                },
                "path": {
                    "type": "string"
                }
            }
        },
        "config.ServerConfig": {
            "type": "object",
            "properties": {
                "host": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "readTimeout": {
                    "type": "string"
                },
                "shutdownTimeout": {
                    "type": "string"
                },
                "writeTimeout": {
                    "type": "string"
                }
            }
        },
        "config.SwaggerConfig": {
            "type": "object",
            "properties": {
                "enable": {
                    "type": "boolean"
                },
                "path": {
                    "type": "string"
                }
            }
        },
        "config.TelemetryConfig": {
            "type": "object",
            "properties": {
                "exporter": {
                    "type": "string"
                },
                "insecure": {
                    "type": "boolean"
                },
                "otlpendpoint": {
                    "type": "string"
                },
                "samplingPercent": {
                    "type": "integer"
                },
                "serviceName": {
                    "type": "string"
                }
            }
        },
        "model.HealthResponse": {
            "type": "object",
            "properties": {
                "service": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "model.ResponseBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "model.StringDataResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "model.StringListResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "model.VersionDataResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/model.VersionResponse"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "model.VersionResponse": {
            "type": "object",
            "properties": {
                "BuildTime": {
                    "type": "string"
                },
                "GitHash": {
                    "type": "string"
                },
                "GoVersion": {
                    "type": "string"
                },
                "Version": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}